refactor routing policy to its own section
